Refactor thread map into modular components
